Fix: Recursive alias definition
The first word of an alias expansion is also checked to see if it
already an alias then expanded except when the first word and the name
of the alias are the same (ie. `alias foo='foo bar'`).

This means cases like:
```
alias oof='foo baz'
alias foo='foo bar'
```
means that `oof` will expand to `foo bar baz` (irrespective of the order
in which the aliases are defined).

To fix the issue either quote the command or escape it.